implemented Add TOTP

+++ src/frontend/src/lib/utils/hex.utils.ts
@@ -1,5 +1,5 @@
 export const hex_decode = (hexString: string) =>
 	Uint8Array.from((hexString.match(/.{1,2}/g) as string[]).map((byte) => parseInt(byte, 16)));
 
-export const hex_encode = (bytes: Uint16Array) =>
+export const hex_encode = (bytes: Uint8Array) =>
 	bytes.reduce((str, byte) => str + byte.toString(16).padStart(2, '0'), '');
